requirements.txt

Django==3.2.6
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-tables2==2.4.0
Faker==8.12.0
